Rename the package and the group, since Sonatype wouldn't let me publish this otherwise.
Jonathan Cobb has not attended the project for 9 years, so I consider it abandoned and ready for a takeover :)
