+++ pom.xml
@@ -2,15 +2,15 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <name>YAML Merge</name>
+    <name>YAML Merger</name>
 
-    <groupId>ch.zizka.yaml.merge</groupId>
-    <artifactId>yaml-merge</artifactId>
+    <groupId>ch.zizka.yaml.merger</groupId>
+    <artifactId>yaml-merger</artifactId>
     <version>1.0.2-SNAPSHOT</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <yamlMerge.mainClass>org.cobbzilla.util.yml.main.MergeYml</yamlMerge.mainClass>
+        <yamlMerge.mainClass>ch.zizka.yaml.merger.main.MergeYaml</yamlMerge.mainClass>
     </properties>
 
     <licenses>
@@ -32,9 +32,7 @@
             <organization>Ondrej Zizka</organization>
             <organizationUrl>https://www.zizka.ch/</organizationUrl>
         </developer>
-        <developer>
-            <name>Jonathan Cobb</name>
-        </developer>
+        <developer><name>Jonathan Cobb</name></developer>
     </developers>
 
     <scm>
@@ -66,7 +64,7 @@
 
         <!-- mustache templates -->
         <dependency>
-            <groupId>com.github.spullara.mustache.java</groupId><artifactId>compiler</artifactId><version>0.9.11</version>
+            <groupId>com.github.spullara.mustache.java</groupId><artifactId>compiler</artifactId><version>0.9.10</version>
             <exclusions>
                 <!-- <exclusion> <groupId>com.google.guava</groupId> <artifactId>guava</artifactId> </exclusion> -->
             </exclusions>
@@ -231,8 +229,11 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId><artifactId>maven-compiler-plugin</artifactId><version>3.11.0</version>
                 <configuration>
+                    <!--
                     <source>11</source>
                     <target>11</target>
+                    -->
+                    <release>11</release>
                     <showWarnings>true</showWarnings>
                 </configuration>
             </plugin>
