+++ src/test/java/ch/zizka/yaml/merger/YamlMergerTest.java
@@ -1,4 +1,4 @@
-package org.cobbzilla.util.yml;
+package ch.zizka.yaml.merger;
 
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -12,13 +12,9 @@ import java.util.Map;
 
 import static org.junit.Assert.*;
 
-/**
- * (c) Copyright 2013 Jonathan Cobb
- * This code is available under the Apache License, version 2: http://www.apache.org/licenses/LICENSE-2.0.html
- */
-public class YmlMergerTest {
+public class YamlMergerTest {
 
-    private static final Logger LOG = LoggerFactory.getLogger(YmlMergerTest.class);
+    private static final Logger LOG = LoggerFactory.getLogger(YamlMergerTest.class);
 
     public static final String YAML_1 = getResourceFile("test1.yaml");
     public static final String YAML_2 = getResourceFile("test2.yaml");
@@ -29,9 +25,9 @@ public class YmlMergerTest {
 
 
     private final Yaml yaml = new Yaml();
-    private final YmlMerger merger = new YmlMerger();
+    private final YamlMerger merger = new YamlMerger();
 
-    @SuppressWarnings({ "deprecation", "unchecked" })
+    @SuppressWarnings({ "unchecked" })
 	@Test
     public void testMerge2Files () throws Exception {
         final Map<String, Object> merged = merger.mergeYamlFiles(new String[]{YAML_1, YAML_2});
@@ -43,7 +39,7 @@ public class YmlMergerTest {
         final String mergedYmlString = merger.exportToString(merged);
         LOG.info("Resulting YAML: \n"+ mergedYmlString);
 
-        final Map<String, Object> reloadedYaml = (Map<String, Object>) yaml.load(mergedYmlString);
+        final Map<String, Object> reloadedYaml = yaml.load(mergedYmlString);
         dbconfig = (Map<String, Object>) reloadedYaml.get("database");
         assertEquals("wrong user", dbconfig.get("user"), "alternate-user");
         assertEquals("wrong db url", dbconfig.get("url"), "jdbc:mysql://localhost:3306/some-db");
@@ -51,7 +47,7 @@ public class YmlMergerTest {
         assertEquals("wrong db url", dbProperties.get("hibernate.dialect"), "org.hibernate.dialect.MySQL5InnoDBDialect");
     }
 
-    @SuppressWarnings({ "deprecation", "unchecked" })
+    @SuppressWarnings({ "unchecked" })
 	@Test
     public void testMergeFileIntoSelf () throws Exception {
         final Map<String, Object> merged = merger.mergeYamlFiles(new String[]{YAML_1, YAML_1});
@@ -60,26 +56,25 @@ public class YmlMergerTest {
         assertEquals("wrong db url", dbconfig.get("url"), "jdbc:mysql://localhost:3306/some-db");
     }
 
-    @SuppressWarnings("deprecation")
-	@Test
+    @Test
     public void testNullValue () throws Exception {
         final Map<String, Object> merged = merger.mergeYamlFiles(new String[]{YAML_NULL});
         assertNotNull(merged.get("prop1"));
         assertNull(merged.get("prop2"));
     }
 
-    @SuppressWarnings({ "deprecation", "unchecked" })
+    @SuppressWarnings({"unchecked"})
 	@Test
     public void testSubstitutionValueWithColon () throws Exception {
         Map<String, String> variables = Collections.singletonMap("ENV_VAR", "localhost");
-        final Map<String, Object> merged = new YmlMerger().setVariablesToReplace(variables).mergeYamlFiles(new String[]{YAML_COLON});
+        final Map<String, Object> merged = new YamlMerger().setVariablesToReplace(variables).mergeYamlFiles(new String[]{YAML_COLON});
         final Map<String, Object> hash = (Map<String, Object>) merged.get("memcache");
         assertEquals(hash.get("one_key"), "value1");
         assertEquals(hash.get("another_key"), "localhost:22133");
         assertEquals(hash.get("some_other_key"), "value2");
     }
 
-    @SuppressWarnings({ "unchecked", "deprecation"})
+    @SuppressWarnings({"unchecked"})
 	@Test
     public void testMerge2Lists () throws Exception {
         final Map<String, Object> merged = merger.mergeYamlFiles(new String[]{MERGE_YAML_1, MERGE_YAML_2});
